backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-ollama
langchain-qdrant
qdrant-client
pypdf
openai
python-dotenv
python-multipart
